NFP-4 adding proper webpack config + sass loader
