+++ package.json
@@ -1,17 +1,26 @@
 {
   "scripts": {
-    "build": "webpack --mode production --config web/webpack.config.js"
+    "clean": "rm -rf *.html css/ js/",
+    "build": "npm run clean && npm i && webpack --config web/webpack.prod.config.js",
+    "dev": "npm i && webpack-dev-server --config web/webpack.dev.config.js"
   },
   "dependencies": {
     "clean-webpack-plugin": "^0.1.19",
-    "file-loader": "^2.0.0",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
     "knex": "^0.15.2",
+    "mini-css-extract-plugin": "^0.4.2",
+    "node-sass": "^4.9.3",
     "pug-html-loader": "^1.1.5",
     "ramda": "^0.25.0",
+    "sass-loader": "^7.1.0",
     "sqlite3": "^4.0.2",
     "webpack": "^4.17.1",
-    "webpack-cli": "^3.1.0"
+    "webpack-cli": "^3.1.0",
+    "css-loader": "^1.0.0",
+    "style-loader": "^0.23.0"
+  },
+  "devDependencies": {
+    "webpack-dev-server": "^3.1.7"
   }
 }
