+++ web/src/home.js
@@ -1,0 +1,1 @@
+import '../views/home/scss/home.scss';
