+++ web/views/home/scss/home.scss
@@ -1,0 +1,1 @@
+@import '../../../scss/_reset.scss';
