+++ web/views/layout.pug
@@ -10,4 +10,4 @@ html(lang='pt-br')
     block header
     block content
     block footer
-    block script
+    block scripts
